+++ presets/npm/src/index.ts
@@ -1,129 +1,57 @@
-#!/usr/bin/env node
 import path from 'path'
 import fs from 'fs-extra'
-import minimist from 'minimist'
 import execa from 'execa'
-import { logger } from '@ombro/logger'
 
-import {
-  CustomScriptType,
-  PresetOptions,
-  PresetRunner,
-  VRNBoilerplateConfig,
-} from '@vrn-deco/boilerplate-protocol'
+import type { PresetOptions, PresetRunner } from '@vrn-deco/boilerplate-protocol'
+import { BaseRunner } from '@vrn-deco/boilerplate-preset-base'
 
-const RUNNER_NAME = 'NPMPresetRunner'
+type PMInstallCommand = {
+  [P in PresetOptions['packageManager']]: string
+}
+const installCommand: PMInstallCommand = {
+  npm: 'install',
+  yarn: 'install', // full amount of installed
+  pnpm: 'install',
+}
 
-// implements PresetRunner
-export const runner: PresetRunner = async (opts: Partial<PresetOptions>) => {
-  logger.verbose(`<${RUNNER_NAME}> startup...`)
+export class NPMRunner extends BaseRunner {
+  override async install(): Promise<boolean> {
+    // The custom script is invoked and exits
+    if (await super.install()) return true
 
-  const argv = minimist(process.argv.slice(2))
-  // prepare and params validate...
-  const {
-    targetDir = argv['target-dir'] ?? argv.targetDir ?? process.cwd(),
-    boiPackageDir = argv['boi-package-dir'] ?? argv.boiPackageDir,
-    name = argv.name,
-    version = argv.version,
-    author = argv.author,
-  } = opts
+    // update package.json
+    this.updatePackageJSON()
 
-  logger.verbose(`targetDir: ${targetDir}`)
-  logger.verbose(`boiPackageDir: ${boiPackageDir}`)
-  logger.verbose(`name: ${name}`)
-  logger.verbose(`version: ${version}`)
-  logger.verbose(`author: ${author}`)
-
-  if (!targetDir) throw new Error('targetDir is required')
-  if (!boiPackageDir) throw new Error('boiPackageDir is required')
-  if (!name) throw new Error('name is required')
-  if (!version) throw new Error('version is required')
-  if (!author) throw new Error('author is required')
-
-  fs.mkdirpSync(targetDir)
-
-  const boiConfigFile = path.join(boiPackageDir, 'vrn-boilerplate.json')
-  if (!fs.existsSync(boiConfigFile)) throw new Error('vrn-boilerplate.json is not exists.')
-  const boiConfig: VRNBoilerplateConfig = fs.readJsonSync(boiConfigFile)
-
-  const boiDir = path.resolve(boiPackageDir, 'boilerplate')
-  logger.verbose(`boiDir: ${boiDir}`)
-  if (!fs.pathExistsSync(boiDir)) throw new Error('boilerplate directory is not exists.')
-
-  // execute preset handler or custom script
-  await initHandler()
-  await installHandler()
-  await cleanHandler()
-
-  logger.verbose(`<${RUNNER_NAME}> finished.`)
-
-  /**
-   * init step
-   */
-  async function initHandler() {
-    logger.verbose(`<${RUNNER_NAME}> -> initHandler`)
-    // using custom script...
-    if (boiConfig.init) {
-      execCustomScript(boiConfig.init.scriptType, boiConfig.init.scriptFile)
-      return
+    // install dependencies
+    const { autoInstallDeps = false } = this.options
+    if (autoInstallDeps) {
+      await this.installDeps()
     }
 
-    // do something...
+    return false
   }
 
-  /**
-   * install step
-   */
-  async function installHandler() {
-    logger.verbose(`<${RUNNER_NAME}> -> installHandler`)
-    // using custom script...
-    if (boiConfig.install) {
-      execCustomScript(boiConfig.install.scriptType, boiConfig.install.scriptFile)
-      return
-    }
-
-    // copy boilerplate files to targetDir...
-    fs.copySync(boiDir, targetDir)
-
-    // do something...
-    const pkgFile = path.join(targetDir, 'package.json')
+  private updatePackageJSON() {
+    const { name, version, author } = this
+    const pkgFile = path.join(this.targetDir, 'package.json')
     const pkg = fs.readJsonSync(pkgFile)
     fs.writeJsonSync(pkgFile, { ...pkg, name, version, author }, { spaces: 2 })
   }
 
-  /**
-   * clean step
-   */
-  async function cleanHandler() {
-    logger.verbose(`<${RUNNER_NAME}> -> cleanHandler`)
-    // using custom script...
-    if (boiConfig.clean) {
-      execCustomScript(boiConfig.clean.scriptType, boiConfig.clean.scriptFile)
-      return
-    }
-
-    // do something...
-  }
-
-  function execCustomScript(scriptType: CustomScriptType, scriptFile: string) {
-    const scriptInterpreter: Record<CustomScriptType, string> = {
-      sh: 'bash',
-      js: 'node',
-      py: 'python3',
-    }
-    const absoluteScriptFile = path.join(boiDir, scriptFile)
-    logger.verbose(`exec custom script: ${absoluteScriptFile}`)
-    execa.sync(
-      scriptInterpreter[scriptType] ?? scriptInterpreter.sh,
-      [absoluteScriptFile, `--name`, name, `--version`, version, `--author`, author],
-      {
-        stdio: 'inherit',
-        cwd: targetDir,
-      },
-    )
+  private async installDeps(): Promise<void> {
+    const { targetDir, packageManager = 'npm' } = this.options
+    await execa(packageManager, [installCommand[packageManager]], {
+      cwd: targetDir,
+      stdio: 'inherit',
+    })
   }
 }
 
+export const runner: PresetRunner = async (options) => {
+  await new NPMRunner(options).run()
+}
+
+// Direct run
 if (require.main === module) {
   runner({})
 }